Move Match Options to separate component
