+++ quick-start/src/main/ui/app/components/flows-new/edit-flow/mastering/ui/matching-ui.component.scss
@@ -1,179 +1,3 @@
 .matching {
   padding: 10px 40px;
 }
-
-.otherLabel {
-  display: inline-block;
-  width: 160px;
-}
-
-$header-height: 100px !default;
-
-button {
-  outline: none;
-}
-
-.page-layout {
-  position: relative;
-  overflow: hidden;
-
-  // Carded layout
-  &.carded {
-    display: flex;
-    flex-direction: column;
-    flex: 1 0 auto;
-    width: 100%;
-    min-width: 100%;
-
-    // Top bg
-    > .top-bg {
-      position: absolute;
-      z-index: 1;
-      top: 0;
-      right: 0;
-      left: 0;
-    }
-
-    // Fullwidth
-    &.fullwidth {
-
-      // Center
-      > .center {
-        display: flex;
-        flex-direction: column;
-        flex: 1 0 auto;
-        position: relative;
-        z-index: 2;
-        padding: 0 32px;
-        width: 100%;
-        min-width: 0;
-        max-width: 100%;
-        height: 100%;
-        max-height: 100%;
-
-        > .header {
-          height: $header-height !important;
-          min-height: $header-height !important;
-          max-height: $header-height!important;
-        }
-
-        > .content-card {
-          display: flex;
-          flex-direction: column;
-          flex: 1 0 auto;
-          overflow: hidden;
-
-          > .content {
-            flex: 1 0 auto;
-          }
-        }
-      }
-    }
-  }
-}
-
-.matching-table {
-  flex: 1 1 auto;
-  border-bottom: 1px solid rgba(0, 0, 0, .12);
-  overflow: auto;
-  -webkit-overflow-scrolling: touch;
-
-  a {
-    color: black;
-  }
-
-  a:hover {
-    text-decoration-color: blue;
-  }
-
-  .mat-header-row {
-    min-height: 64px;
-  }
-
-  .flow {
-    position: relative;
-    cursor: pointer;
-    height: 84px;
-  }
-
-  .mat-cell {
-    min-width: 0;
-    display: flex;
-    align-items: center;
-  }
-
-  .mat-column-propertyName {
-    flex: 0 1 200px;
-  }
-
-  .mat-column-matchType {
-    flex: 0 1 180px;
-  }
-
-  .mat-column-weight {
-    flex: 0 1 100px;
-  }
-
-  .mat-column-other {
-    flex: 0 1 480px;
-  }
-
-  .mat-column-actions {
-    flex: 0 1 80px;
-  }
-
-  .mat-header-cell .mat-icon {
-    cursor: pointer;
-  }
-
-  .logo {
-    padding-right: 30px;
-  }
-
-  .run-flow-button {
-    margin: 0 15px 0 7px;
-  }
-
-  .actions-menu-icon {
-    margin-left: 10px;
-  }
-
-  .quantity-indicator {
-    display: inline-block;
-    vertical-align: middle;
-    width: 8px;
-    height: 8px;
-    border-radius: 4px;
-    margin-right: 8px;
-
-    & + span {
-      display: inline-block;
-      vertical-align: middle;
-    }
-  }
-
-  .active-icon {
-    border-radius: 50%;
-  }
-}
-
-.fill-space {
-  flex: 1 1 auto;
-}
-
-.name-container {
-  position: relative;
-  width: 240px;
-}
-
-.logo {
-  padding-right: 30px;
-}
-
-.mat-column-actions .mat-icon {
-  cursor: pointer;
-}
-
-.capitalize {
-   text-transform: capitalize;
-}
